requirements.txt
# Core ML — no strict sklearn pin so your installed version is used for training
scikit-learn>=1.1.0
pandas>=1.5.0
numpy>=1.23.0
joblib>=1.2.0
PyYAML>=6.0
packaging>=21.0

# Experiment tracking
mlflow>=2.0.0

# Serving
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# UI
streamlit>=1.28.0

# Testing
pytest>=7.0.0
httpx>=0.24.0
